separate link dialog from text format bar
