+++ src/css/_styles.less
@@ -25,5 +25,5 @@
 @import "settings-dialog";
 @import "stage";
 @import "alertify";
-@import "link-editor";
+@import "link-dialog";
 @import "text-format-bar";
